ARM: dts: rk356x: align bootsource_instance numbering and MMC DT aliases
A common scripting pattern in barebox is continuing boot from
the current bootsource:

  if [ "$bootsource" = "mmc" ]; then
	global.boot.default="mmc${bootsource_instance} ${global.boot.default}"
  fi

This fails on the RK356x boards, because of a mismatch in numbering:
block device numbering is dictated by the mmc* DT aliases and numbering
used for bootsource_instance is taken as is from what the bootrom reports.

On SoCs like i.MX, this is a much rarer issue, because the upstream DT
has predefined aliases in the same order as reported by the bootrom.

In Rockchip DTs on the other hand, aliases are delegated to the board
device trees and all boards we support have aliases in an order
different from what the bootrom numbering.

Fix this by employing the newly supported barebox,bootsource-* mapping
table. As a result $bootsource_instance will change. We fix the upstream
code depending on specific numbering here, but custom scripts and board
code may still be broken.

Co-developed-by: Johannes Zink <j.zink@pengutronix.de>
Signed-off-by: Johannes Zink <j.zink@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Tested-by: Johannes Zink <j.zink@pengutronix.de> # Radxa Rockpi3A
Link: https://lore.barebox.org/20220720055042.3510276-4-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>