gpiolib: Add of_xlate support
The of_xlate callback is used to tanslate a device-tree GPIO specifier
into a chip relative GPIO number and flags. This callback needs to be
implemented by GPIO controller driver which doesn't uses the "classic"
device-tree description, which is:
    gpio = <&gpio-phandle gpio_num gpio_flags>;

But uses something else, like this (but not limited to):
    gpio = <&gpio-phandle gpio_bank gpio_num gpio_flags>;

Currently no GPIO driver require this of_xlate callback, but this will
be used for sunxi's GPIO support.

Signed-off-by: Jules Maselbas <jmaselbas@kalray.eu>
Link: https://lore.barebox.org/20220805135928.32725-1-jmaselbas@kalray.eu
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>