common: machine_id: simplify early exit
We don't need the goto if we haven't done anything to clean up anyway.
also globalvar_add_simple("machine_id", NULL) is a no-op when we have
just called globalvar_add_simple above with an actual argument.
It doesn't clean the parameter, nor should it, because the code is
executed for the successful code as well and there is nothing that can
fail that late.

This slightly alters behavior: Whereas before $global.machine_id was
always defined when CONFIG_MACHINE_ID is enabled, it's now only
defined when it's non-empty.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220808062010.390394-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>