ARM: dts: stm32mp: add bootrom serial (DFU) reboot mode
The same register we use for the reboot mode communication between Linux
and barebox is also sampled by the bootrom. If the lowest 16 bits of it
are 0x00FF, the bootrom will enter DFU mode. Add the appropriate
reboot-mode, but with a barebox, prefix, so Linux doesn't automatically
provide it to userspace along with the existing loader and recovery
modes.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220809062938.1750369-4-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>