ARM: i.MX8M: atf: mark ATF start functions as __noreturn
The early exit in imx8m_atf_load_bl31 doesn't serve any real purpose.
If a too big TF-A ends up being loaded we shouldn't return, but
just panic outright, so the user fixes their barebox build.

Change the WARN_ON to a BUG_ON() and then mark all callers of this
function as __noreturn to document their purpose and to guide
optimization.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220816073056.45694-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>