drivers: mtd: nand: nand_mxs: enable nand_mxs on imx7
The i.MX7 GPMI NAND controller is completely backwards compatible with
the version of the IP on the i.MX6. There are additional features like
increased ECC strength, that we may want to use in future, but for now
treating it the same as an i.MX6 works without problems.

Since the gpmi_nand driver requires the MXS_APBH_DMA driver, also switch it on.

Signed-off-by: Johannes Zink <j.zink@pengutronix.de>
Link: https://lore.barebox.org/20220830080613.438124-1-j.zink@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>