i2c: implement of_i2c_register_devices_by_node()
i2c controllers probe the devices on their bus themselves, so in the
case the i2c controller was already probed, we currently have no way to
have it reread the device tree after fixing it up or applying an
overlay. Add a new helper that retriggers the device creation.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220905103613.1750340-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>