ARM: Fix barebox header generation on arm64
When using multiple ENTRY_FUNCTION_WITH_STACK an entry_prologue
is emitted for each usage. The entry_prologues end up in the
same section and thus can't be discarded with the effect that the
barebox magic is at the wrong place. To fix this put the entry_prologues
in function specific sections so that the linker can discard the
unused ones.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220909072902.3260039-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>