ARM: i.MX23: fix memory size calulcation
According to the i.MX233 Reference Manual (i.MX233RM, Rev. 4, 03 April
2009) section "14.1.1 AHB Address Ranges", the formula to calculate the
DRAM memory size is:

  dram_memory_available = 2 * 2^col * 2^row * (# dram_devices) * (# banks_per_device)

The calulcation in barebox misses the first factor, so the result is
only half as big as it should be.

On an iMX233 OLinuXino board the following errors can be observed:

  mmu: Critical Error: Can't request SDRAM region for ttb at 43fe4000
  Error: Cannot request SDRAM region for stack

The faulty calulcation leads to 32 MB being detected (as seen in the
output of `iomem`), although 64 MB should be available (as passed to
barebox_arm_entry() in lowlevel code).

Fix the memory calculation by adding the missing factor 2.

Fixes: 7158c5987e6 ("ARM: i.MX23: Add memory size detection")
Signed-off-by: Bastian Krause <bst@pengutronix.de>
Signed-off-by: Jürgen Borleis <jbe@pengutronix.de>
Link: https://lore.barebox.org/20220909143915.4081275-1-bst@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>