ARM: imx_v7_defconfig: Limit TFTP window size
The FEC ethernet controller is too slow to catch up with a gigabit
link. This hasn't been a problem until now, because the protocols
we use in barebox only have a single packet in flight. With TFTP
window size support and a gigabit link though we get receive
buffer overruns resulting in messages like:

WARNING: eth0: error frame: 0x2feca098 0x00000806

The resulting speed is slower than without TFTP windowsize support due
to the necessary retransmits. Limit the window size to a value that
seems to be working on i.MX6 boards.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>