net: Bring up all interfaces when going interactive
So far we only bring up network interfaces when we actually need them.
This means we could be idling in the shell for long and once the user
decides to do networking he has to wait for the link to be established.
We can do better: Before going interactive bring up all known network
interfaces which makes the links established when the user needs them.

To implement this we have to rework carrier checking a bit, because
otherwise barebox would wait for the links to be established before
continuing.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>