bbu: fix exporting i.MX NAND bbu handler over fastboot
handler->devicefile for i.MX NAND is nand0.barebox, which lacks a /dev/
prefix. This is ok for detection, as device_detect_by_name expects the
cdev name and devpath_to_name() will strip a /dev/ if available.

For stat() however, we need to add back /dev/, otherwise the file can't
be found. Rework the code to do that.

This fixes an issue where usbgadget -A '' -b would not export a i.MX
NAND barebox update handler.

Fixes: 726a802456bc ("common: bbu: only add available handlers")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220928085622.2677478-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>