Revert "lib: zstd: sync with Linux"
This was part of a larger series that made it possible to use other
compression algos than LZO for device tree and to use zstd more widely
for barebox and kernel compression. The latter zstd-related parts needed
rework and were not merged, but the zstd update was applied. It broken
in turn the UBIFS zstd support. As we get nothing out of the update
without the not applied commits that need rework anyway, revert this
commit.

This reverts commit b4a9782d4f56333e897dccc35c2c27e2605f6b93.

Reported-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220929102537.1767458-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>