ARM: i.MX7: replace hardcoded UART clocking defines
We currently have the clock defines for 1-3, but lack 4-6.
Add generic defines that can be used for all of 1-6 and
start using them in the header.

The old defines are not used outside the file, so drop them.
Out-of-tree users can just move the number into the parenthesis:

	IMX7_CCM_CCGR_UART1 -> IMX7_CCM_CCGR_UART(1)
	IMX7_UART1_CLK_ROOT ->  IMX7_UART_CLK_ROOT(1)

Consulting the data sheet also showed that IMX7_UART_CLK_ROOT__OSC_24M
is the same value for all UARTs, so we omit the argument there.

No functional change.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20221017070702.1457936-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>